+++ tests/integration/test_model_instance.py
@@ -16,6 +16,7 @@ from gamspy import (
     Alias,
     Container,
     Equation,
+    FreezeOptions,
     Model,
     ModelInstanceOptions,
     ModelStatus,
@@ -346,14 +347,21 @@ def test_validations(data):
 
     transport.freeze(modifiables=[x.up], options=Options(lp="conopt"))
 
-    # Test model instance options
+    # Test freeze options
     transport.solve(
         solver="conopt",
-        model_instance_options=ModelInstanceOptions(debug=True),
+        model_instance_options=FreezeOptions(debug=True),
     )
     assert math.isclose(transport.objective_value, 153.675, rel_tol=1e-6)
     assert os.path.exists("dict.txt")
     assert os.path.exists("gams.gms")
+
+    # ModelInstanceOptions should throw a warning
+    with pytest.warns(UserWarning):
+        transport.solve(
+            solver="conopt",
+            model_instance_options=ModelInstanceOptions(debug=True),
+        )
 
     # Test solver options
     with open("_out.txt", "w") as file: