+++ CHANGELOG.md
@@ -6,6 +6,10 @@ GAMSPy 1.9.0
 - General
   - Validate solver options for most of the solvers. It can be disable through VALIDATE_SOLVER_OPTIONS option.
   - get the value of objective estimation and the number of nodes used after frozen solves
+  - Add description argument for model objects.
+  - Make GAMSPy symbols non-iterable.
+  - Inherit output argument from the container in solve function if the output argument is not specified.
+  - Rename ModelInstanceOptions to FreezeOptions and add a warning for the usage of ModelInstanceOptions.
 
 GAMSPy 1.8.0
 ------------